# [Performance][PD Disaggregation] optimize TokenToKVPoolAllocator by sorting free pages