<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

In large-scale distributed inference, especially with PD Disaggregation, fast and efficient management of KV cache token indices is crucial. Fragmentation of available indices can disrupt peer-to-peer (p2p) data transfers by making it harder to allocate large contiguous regions, resulting in inefficient communication patterns. This fragmentation directly increases end-to-end latency, especially the time to first token (TTFT).
This PR introduces release cache mechanisms to optimize TokenToKVPoolAllocator page management. By deferring the actual release of freed pages and supporting batch reclamation, this design significantly reduces allocation contention and fragmentation during high-concurrency operations, thereby boosting p2p transfer efficiency and reducing TTFT.

## Modifications

### Introduced release_pages
Freed token indices are now buffered in a temporary cache rather than immediately added to the free pool.

### On allocation, merging and sorting
When an allocation request cannot be satisfied by the current free pool alone, the allocator will merge the buffered indices from release cache with the free pool, sort them, and then retry the allocation. This sorting and merging process helps to coalesce adjacent free regions, mitigating fragmentation.

### Test on 2 * A10 * 8（RDMA），qwen 0.6b，1P1D

Case 1（200 warmup；400 nr；16 concurrency; 1024 input len; 512 output len）
without optimization:  TTFT 81ms
with optimization: TTFT 69ms（improve 14.8%）

Case 2（200 warmup；400 nr；16 concurrency; 2048 input len; 1024 output len）
without optimization:  TTFT 144ms
with optimization: TTFT 108ms（improve 25%）

A simple comparison of the number of kv_blocks that need to be transferred by send_kvcache() (1024 in; 512 out)
<img width="4340" height="2236" alt="sglang" src="https://github.com/user-attachments/assets/dd9cbd75-8105-4dbd-a0d6-79441172affa" />

### Detailed data
#### 1024 input len; 512 output len without optimization
Backend:                                 sglang
Traffic request rate:                    inf
Max request concurrency:                 16
Successful requests:                     400
Benchmark duration (s):                  188.20
Total input tokens:                      409600
Total generated tokens:                  204800
Total generated tokens (retokenized):    204796
Request throughput (req/s):              2.13
Input token throughput (tok/s):          2176.35
Output token throughput (tok/s):         1088.18
Total token throughput (tok/s):          3264.53
Concurrency:                             15.98
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   7517.90
Median E2E Latency (ms):                 7501.12
---------------Time to First Token----------------
Mean TTFT (ms):                          81.29
Median TTFT (ms):                        74.55
P99 TTFT (ms):                           300.72
---------------Inter-Token Latency----------------
Mean ITL (ms):                           14.55
Median ITL (ms):                         14.47
P95 ITL (ms):                            14.88
P99 ITL (ms):                            15.69
Max ITL (ms):                            33.38

#### 1024 input len; 512 output len with optimization
Backend:                                 sglang
Traffic request rate:                    inf
Max request concurrency:                 16
Successful requests:                     400
Benchmark duration (s):                  185.05
Total input tokens:                      409600
Total generated tokens:                  204800
Total generated tokens (retokenized):    204792
Request throughput (req/s):              2.16
Input token throughput (tok/s):          2213.43
Output token throughput (tok/s):         1106.71
Total token throughput (tok/s):          3320.14
Concurrency:                             15.98
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   7393.70
Median E2E Latency (ms):                 7386.50
---------------Time to First Token----------------
Mean TTFT (ms):                          69.35
Median TTFT (ms):                        59.20
P99 TTFT (ms):                           363.33
---------------Inter-Token Latency----------------
Mean ITL (ms):                           14.33
Median ITL (ms):                         14.27
P95 ITL (ms):                            14.67
P99 ITL (ms):                            15.65
Max ITL (ms):                            35.02

#### 2048 input len; 1024 output len without optimization
Backend:                                 sglang
Traffic request rate:                    inf
Max request concurrency:                 16
Successful requests:                     400
Benchmark duration (s):                  395.88
Total input tokens:                      819200
Total generated tokens:                  409600
Total generated tokens (retokenized):    409584
Request throughput (req/s):              1.01
Input token throughput (tok/s):          2069.32
Output token throughput (tok/s):         1034.66
Total token throughput (tok/s):          3103.98
Concurrency:                             15.97
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   15803.31
Median E2E Latency (ms):                 15800.76
---------------Time to First Token----------------
Mean TTFT (ms):                          144.02
Median TTFT (ms):                        118.62
P99 TTFT (ms):                           975.52
---------------Inter-Token Latency----------------
Mean ITL (ms):                           15.31
Median ITL (ms):                         15.17
P95 ITL (ms):                            16.62
P99 ITL (ms):                            16.91
Max ITL (ms):                            44.18

#### 2048 input len; 1024 output len with optimization
Backend:                                 sglang
Traffic request rate:                    inf
Max request concurrency:                 16
Successful requests:                     400
Benchmark duration (s):                  392.21
Total input tokens:                      819200
Total generated tokens:                  409600
Total generated tokens (retokenized):    409586
Request throughput (req/s):              1.02
Input token throughput (tok/s):          2088.66
Output token throughput (tok/s):         1044.33
Total token throughput (tok/s):          3132.99
Concurrency:                             15.98
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   15669.23
Median E2E Latency (ms):                 15658.25
---------------Time to First Token----------------
Mean TTFT (ms):                          108.04
Median TTFT (ms):                        89.39
P99 TTFT (ms):                           733.34
---------------Inter-Token Latency----------------
Mean ITL (ms):                           15.21
Median ITL (ms):                         15.08
P95 ITL (ms):                            16.57
P99 ITL (ms):                            16.89
Max ITL (ms):                            39.90

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.