@YiXR 2025-07-23
I've just fix a problem when using alloc_extend() and alloc_decode(), please help review this, thx. @xiezhq-hermann @ShangmingCai